refactor: move `enrichers` to `helper` module
