+++ module/hooks/init.mjs
@@ -14,7 +14,7 @@ import * as canvas from "../canvas/_module.mjs";
 import * as data from "../data/_module.mjs";
 import * as dice from "../dice/_module.mjs";
 import * as documents from "../documents/_module.mjs";
-import * as enrichers from "../enrichers.mjs";
+import * as enrichers from "../helper/enrichers.mjs";
 
 import { staticStatusId } from "../helper/document.mjs";
